start fixing std::optional implementation
std::optional must be trivially destructible if T is trivially
destructible, for this there must not be a Destructor.
This is implemented by using a __optional_base class that only has a
destructor if it is not trivially destructible.

The base class also implements std::optional::reset(), since the
reset() method has to destroy the contained object.